+++ server/routers/authRoute.js
@@ -7,8 +7,6 @@ const {
   forgotPassword,
   resetPassword,
 } = require("../controller/authController");
-const { sendMailsFromSheet } = require("../controller/authController.js");
-
 
 const cors = require("cors");
 
@@ -29,5 +27,5 @@ router.post("/forgot-password", forgotPassword);
 
 router.get("/reset-password/:id/:token", resetPassword);
 
-router.post("/send-mail", sendMailsFromSheet);
+
 module.exports = router;
